+++ leeger/model/abstract/UniqueId.py
@@ -17,7 +17,7 @@ class UniqueId(ABC):
         return self.__id
 
     @id.setter
-    def id(self, value: str, warnUser=True):
+    def id(self, value: str):
         self.__id = value
 
     @id.deleter