+++ leeger/model/league/Owner.py
@@ -3,11 +3,12 @@ from __future__ import annotations
 from dataclasses import dataclass
 
 from leeger.model.abstract.UniqueId import UniqueId
+from leeger.util.JSONDeserializable import JSONDeserializable
 from leeger.util.JSONSerializable import JSONSerializable
 
 
 @dataclass(kw_only=True, eq=False)
-class Owner(UniqueId, JSONSerializable):
+class Owner(UniqueId, JSONSerializable, JSONDeserializable):
     name: str
 
     def __eq__(self, otherOwner: Owner) -> bool:
@@ -22,3 +23,9 @@ class Owner(UniqueId, JSONSerializable):
             "id": self.id,
             "name": self.name
         }
+
+    @staticmethod
+    def fromJson(d: dict) -> Owner:
+        owner = Owner(name=d["name"])
+        owner.id = d["id"]
+        return owner