+++ leeger/model/league/YearSettings.py
@@ -1,13 +1,19 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
+from typing import Optional
 
+from leeger.util.JSONDeserializable import JSONDeserializable
 from leeger.util.JSONSerializable import JSONSerializable
 
 
 @dataclass(kw_only=True, eq=False)
-class YearSettings(JSONSerializable):
-    leagueMedianGames: bool = False
+class YearSettings(JSONSerializable, JSONDeserializable):
+    leagueMedianGames: Optional[bool] = False
+
+    def __post_init__(self):
+        if self.leagueMedianGames is None:
+            self.leagueMedianGames = False
 
     def __eq__(self, otherYearSettings: YearSettings) -> bool:
         """
@@ -19,3 +25,7 @@ class YearSettings(JSONSerializable):
         return {
             "leagueMedianGames": self.leagueMedianGames
         }
+
+    @staticmethod
+    def fromJson(d: dict) -> YearSettings:
+        return YearSettings(leagueMedianGames=d.get("leagueMedianGames"))