+++ test/test_validate/test_yearSettingsValidation.py
@@ -1,6 +1,5 @@
 import unittest
 
-from leeger.exception import InvalidYearSettingsFormatException
 from leeger.model.league import YearSettings
 from leeger.validate import yearSettingsValidation
 
@@ -10,8 +9,10 @@ class TestYearSettingsValidation(unittest.TestCase):
     TYPE CHECK TESTS
     """
 
-    def test_checkAllTypes_leagueMedianGamesIsntTypeBool_raisesException(self):
-        with self.assertRaises(InvalidYearSettingsFormatException) as context:
-            yearSettings = YearSettings(leagueMedianGames=None)
-            yearSettingsValidation.checkAllTypes(yearSettings)
-        self.assertEqual("leagueMedianGames must be type 'bool'.", str(context.exception))
+    def test_runAllChecks(self):
+        yearSettings = YearSettings(leagueMedianGames=None)
+        yearSettingsValidation.runAllChecks(yearSettings)
+        yearSettings = YearSettings(leagueMedianGames=True)
+        yearSettingsValidation.runAllChecks(yearSettings)
+        yearSettings = YearSettings(leagueMedianGames=False)
+        yearSettingsValidation.runAllChecks(yearSettings)